+++ packages/dartvel_core/lib/http.dart
@@ -10,6 +10,6 @@
 library dartvel_core.http;
 
 export 'src/http/peer_address.dart' show DVIpAddress, DVPeerAddress;
-export 'src/http/router.dart' show Router;
+export 'src/http/router.dart' show DVRouteBodyLimit, Router;
 export 'src/http/wintercg.dart'
     show Request, Response, Headers, Body, URLPattern;